+++ .eslintrc
@@ -4,7 +4,8 @@
     "plugin:react/recommended"
   ],
   "rules": {
-    "space-before-function-paren": 0
+    "space-before-function-paren": 0,
+    "object-curly-spacing": 0
   },
   "env": {
     "browser": true,
