+++ js/config.js
@@ -1,9 +1,5 @@
-'use strict';
-
-var config = {
-    'overpassBase': '//overpass-cfn-production.tilestream.net/api/interpreter',
-    'osmBase': '//www.openstreetmap.org/api/0.6/',
-    'mapboxAccessToken': 'pk.eyJ1IjoicmFzYWd5IiwiYSI6ImNpejVrMjc4eTAwNGczM2thNWozYnJ1OHkifQ.yFRr3Sd39TJiwEguQpIkWQ'
+export const config = {
+    overpassBase: '//overpass-cfn-production.tilestream.net/api/interpreter',
+    osmBase: '//www.openstreetmap.org/api/0.6/',
+    mapboxAccessToken: 'pk.eyJ1IjoicmFzYWd5IiwiYSI6ImNpejVrMjc4eTAwNGczM2thNWozYnJ1OHkifQ.yFRr3Sd39TJiwEguQpIkWQ'
 };
-
-module.exports = config;
