--- lib/index.js
@@ -1,5 +1,0 @@
-export { config } from './config';
-export { getChangeset } from './getChangeset';
-export { query } from './query';
-export { propsDiff } from './propsDiff';
-export { render } from './render';
